backend/requirements.txt
flask
flask-cors
pymupdf
pillow
